rename: RecentSearchView를 SearchView로 변경
